symbol nearly matched

+++ src/xdk/LIBCMT/stddef.h
@@ -7,7 +7,7 @@
 extern "C" {
 #endif
 
-typedef size_t ptrdiff_t;
+typedef int ptrdiff_t;
 #define offsetof(T, mem) ((int)&((T *)0->mem))
 
 #ifdef __cplusplus
